Fix sqlalchemy pool noise and clean up filterwarnings
Move the suppress_sqlalchemy_pool_noise autouse fixture from
io_managers_test.py into a new tests/unit/dagster/conftest.py so it
covers all Dagster unit tests. The same Dagster + Python 3.13 pathlib
teardown cascade that was silenced for io_managers tests was also
appearing in ferceqr_deployment_test.py.

Remove five obsolete filterwarnings entries from pyproject.toml:
- distutils DeprecationWarning (distutils removed in Python 3.12)
- open_binary / open_text DeprecationWarning (no longer used in
  pudl.glue.ferc1_eia)
- "In a future version" FutureWarning from pudl.helpers (fixed by
  the resample(include_groups=False) change in the prior commit)
- Subclassing validator classes from tableschema (duplicate; covered
  by the broader entry above it)
