+++ tests/unit/dagster/io_managers_test.py
@@ -100,27 +100,6 @@ def test_pkg() -> Package:
     )
 
 
-@pytest.fixture(autouse=True)
-def suppress_sqlalchemy_pool_noise():
-    """Suppress SQLAlchemy NullPool teardown errors from Dagster's ephemeral instance.
-
-    build_output_context() creates a Dagster ephemeral instance backed by its own
-    internal SQLite database. When an OutputContext is GC'd after a failed
-    handle_output(), Dagster's teardown code accesses pathlib internals (_str, _drv)
-    that were removed in Python 3.13, causing a cascade: GeneratorExit propagates
-    through build_resources, the finally block tries to roll back an already-closed
-    SQLite connection, and sqlalchemy.pool logs the ProgrammingError at ERROR level.
-    The tests are correct; this is a Dagster + Python 3.13 compatibility issue.
-    Setting the level here (not in a with-block) ensures it's active during GC
-    teardown, which occurs after the test function returns.
-    """
-    sa_pool_logger = logging.getLogger("sqlalchemy.pool")
-    original_level = sa_pool_logger.level
-    sa_pool_logger.setLevel(logging.CRITICAL)
-    yield
-    sa_pool_logger.setLevel(original_level)
-
-
 @pytest.fixture
 def sqlite_io_manager_fixture(tmp_path, test_pkg) -> SqliteIOManager:
     """Create a SqliteIOManager fixture with a simple database schema."""
